Update README.md
This demo shows an application of distributed multi-clock real-time system modeling based on time-triggered mode, the master-slave cars’ system. In the master-slave cars’ system, it is necessary to ensure that the slave car keeps tracking within a certain distance while the master car runs. In the normal operation mode, the system checks the distance between two cars periodically. The behavior of distance checking is period time-triggered. Once the distance exceeds the pre-set maximum, the slave car will send an alarm message to the master car. After the master car receives the alarm information, it will switch the mode to stop, waiting in place. At the same time, the master car will send the pose information to the slave car. The slave car chases according to the infrared signal and the information fed back from the master car. After the distance is detected to be in the preset value interval, the slave car will send an un-warning message to the master car. After receiving the un-warning signal, the master car will switch back to normal mode and continue to run.